requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.1
django-stdimage==6.0.1
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2023.2
sqlparse==0.4.3
